Move edit menu to new file
